+++ src/test/java/org/cicirello/search/concurrent/ParallelMultistarterSingleThreadTests.java
@@ -117,7 +117,8 @@ public class ParallelMultistarterSingleThreadTests {
 		heur = new TestRestartedMetaheuristic();
 		restarter = new ParallelMultistarter<TestObject>(heur, 1, 1);
 		expected = restarter.getTotalRunLength();
-		restarter.getProgressTracker().setFoundBest();
+		// replaced call to deprecated setFoundBest()
+		restarter.getProgressTracker().update(0, new TestObject(), true);
 		restarter.optimize(1);
 		assertEquals(expected, restarter.getTotalRunLength());
 		restarter.close();
@@ -897,7 +898,8 @@ public class ParallelMultistarterSingleThreadTests {
 					if (elapsed >= stopAtEval) {
 						elapsed = stopAtEval;
 						tracker.stop();
-						tracker.setFoundBest();
+						// Replaces old call to deprecated setFoundBest()
+						tracker.update(1, new TestObject(), true);   
 						c = 1;
 					}
 					break;
@@ -910,7 +912,8 @@ public class ParallelMultistarterSingleThreadTests {
 				case 2:
 					if (elapsed >= findBestAtEval) {
 						elapsed = findBestAtEval;
-						tracker.setFoundBest();
+						// Replaces old call to deprecated setFoundBest()
+						tracker.update(1, new TestObject(), true);
 						c = 1;
 					}
 					break;